add angular view

+++ YouBikeAPI/Services/BikeRepository.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
@@ -9,6 +10,7 @@ using YouBikeAPI.Data;
 using YouBikeAPI.Dtos;
 using YouBikeAPI.Helper;
 using YouBikeAPI.Models;
+using YouBikeAPI.Utilities;
 
 namespace YouBikeAPI.Services
 {
@@ -36,20 +38,56 @@ namespace YouBikeAPI.Services
             _configuration = configuration;
         }
 
-        public async Task<PaginationList<Bike, BikeDto>> GetBikes(int pageNum, int pageSize)
+        public async Task<PaginationList<Bike, BikeDto>> GetBikes(int pageNum, int pageSize, string query)
         {
-            IQueryable<Bike> bikes = _context.Bikes.AsQueryable().OrderBy(b => b.Id);
+            IQueryable<Bike> bikes = _context.Bikes
+                .AsQueryable()
+                .Include(b => b.BikeStation)
+                .OrderBy(b => b.Id);
+
+            if (!string.IsNullOrEmpty(query))
+            {
+                var properyList = new Dictionary<string, Object>();
+                var propertyNames = new List<string> { "Id", "Mileage", "Revenue", "BikeStation.StationName" };
+
+                query = query.Substring(0, 1).ToUpper() + query.Substring(1);
+                if (Enum.IsDefined(typeof(BikeType), query)) propertyNames.Add("BikeType");
+                if ("閒置中".Contains(query) || "租借中".Contains(query)) propertyNames.Add("Rented");
+
+                foreach (var property in propertyNames)
+                {
+                    if (property.Split(".")[0] == "BikeStation")
+                        properyList.Add(property, typeof(BikeStation).GetProperty("StationName").PropertyType);
+                    else if (property.Split(".")[0] == "BikeType")
+                        properyList.Add(property, typeof(System.Enum));
+                    else
+                        properyList.Add(property, typeof(Bike).GetProperty(property.Split(".")[0]).PropertyType);
+                }
+
+                query = "閒置中".Contains(query) ? "false" : query;
+                query = "租借中".Contains(query) ? "true" : query;
+
+                bikes = bikes.CustomSearch(query, properyList);
+            }
+
             return await PaginationList<Bike, BikeDto>.CreateAsync(pageNum, pageSize, bikes, _mapper);
         }
 
         public async Task<Bike> GetBike(int id)
         {
-            return await _context.Bikes.Include((Bike b) => b.Price).SingleOrDefaultAsync((Bike b) => b.Id == id);
+            Bike bike = await _context.Bikes
+                .Include(b => b.Price)
+                .SingleOrDefaultAsync(b => b.Id == id);
+            return bike;
         }
 
         public async Task<BikeDto> GetBikeForUser(int id)
         {
-            Bike bike = await _context.Bikes.Include((Bike b) => b.Price).Include((Bike b) => b.User).FirstOrDefaultAsync((Bike b) => b.Id == id);
+            Bike bike = await _context.Bikes
+                .Include(b => b.BikeStation)
+                .Include(b => b.Price)
+                .Include(b => b.User)
+                .SingleOrDefaultAsync(b => b.Id == id);
             return _mapper.Map<BikeDto>(bike);
         }
 
@@ -77,14 +115,20 @@ namespace YouBikeAPI.Services
             _mapper.Map(bike, targetBike);
         }
 
-        public async Task DeleteBike(int id)
+        public async Task<(string, bool)> DeleteBike(int id)
         {
             Bike bike = await GetBike(id);
-            _context.Bikes.Remove(bike);
+            if (!bike.Rented)
+                _context.Bikes.Remove(bike);
+            else
+            {
+                return ("租借中車輛不得刪除", false);
+            }
             if (bike.BikeStationId.HasValue)
             {
                 (await _bikeStationRepository.GetBikeStationById(bike.BikeStationId)).BikesInsideParkingLot--;
             }
+            return ("刪除成功", true);
         }
 
         public async Task<bool> BikeExists(int id)
@@ -115,25 +159,35 @@ namespace YouBikeAPI.Services
         {
             ApplicationUser currentUser = await _context.Users.SingleOrDefaultAsync((ApplicationUser u) => u.Id == userId);
             if (currentUser == null || currentUser.Bike != null)
-            {
                 return false;
-            }
+
             Bike bike = await GetBike(id);
             if (bike.Rented)
-            {
                 return false;
-            }
+
             bike.Rented = true;
             bike.UserId = userId;
             currentUser.Bike = bike;
             BikeStation station = await _bikeStationRepository.GetBikeStationById(bike.BikeStationId);
-            currentUser.HistoryRoutes.Add(new HistoryRoute
+
+            var route = new HistoryRoute
             {
                 Source = station.Id,
+                SourceName = station.StationName,
                 ApplicationUserId = userId,
                 BikeId = bike.Id,
                 CurrentRoute = true,
-                BorrowTime = DateTime.UtcNow
+                CreationDate = DateTime.UtcNow,
+                ReturnTime = null
+            };
+
+            await _paidmentService.CreateHistoryRoute(route);
+            if (!await SaveAllAsync()) return false;
+
+            currentUser.HistoryRouteItem.Add(new HistoryRouteItem
+            {
+                ApplicationUserId = currentUser.Id,
+                HistoryRouteId = route.Id
             });
             station.BikesInsideParkingLot--;
             station.AvailableBikes.Remove(bike);
@@ -160,6 +214,7 @@ namespace YouBikeAPI.Services
             {
                 return (false, "查無此車站");
             }
+            historyRoute.DestinationName = stationTo.StationName;
             string lan = $"{stationFrom.Latitude},{stationFrom.Longitude}";
             string lon = $"{stationTo.Latitude},{stationTo.Longitude}";
             (bool, string) paymentResult = await _paidmentService.PayBill(userId, lan, lon);
@@ -177,6 +232,24 @@ namespace YouBikeAPI.Services
             stationTo.BikesInsideParkingLot++;
             stationTo.AvailableBikes.Add(bike);
             return (true, "成功歸還");
+        }
+
+        public async Task<int> GetBikesCount()
+        {
+            return await _context.Bikes.CountAsync();
+        }
+
+        public async Task<int> BikeIncreasedInLastMonth()
+        {
+            return await _context.Bikes
+                .DateBetween(-2, -1)
+                .CountAsync();
+        }
+
+        public async Task<IEnumerable<PriceDto>> GetPricesAsync()
+        {
+            var prices = await _context.Prices.ToListAsync();
+            return _mapper.Map<IEnumerable<PriceDto>>(prices);
         }
     }
 }
